🎨 adds Marketplace Handler
