+++ Helper/Marketplace/ExtrasAndDiscountsHandler.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Pagarme\Pagarme\Helper\Marketplace;
+
+use Pagarme\Pagarme\Helper\Marketplace\MarketplaceHandler;
+
+class ExtrasAndDiscountsHandler extends MarketplaceHandler
+{
+    public function __construct()
+    {
+    }
+}
